@@ -3,31 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@material-ui/core": "^4.12.4",
-    "@mui/icons-material": "^5.4.1",
-    "@mui/lab": "^5.0.0-alpha.66",
-    "@mui/material": "^5.10.10",
-    "@mui/styles": "^5.10.10",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.1.2",
+    "@mui/material": "^7.1.2",
+    "@reduxjs/toolkit": "^2.8.2",
     "bootstrap": "^5.1.3",
-    "csv-parse": "^5.0.4",
     "eslint-config-react-app": "^7.0.0",
-    "fbjs": "^3.0.2",
-    "konva": "^8.3.2",
-    "react": "^17.0.2",
+    "konva": "^8.4.3",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-dropdown": "^1.9.2",
-    "react-konva": "^17.0.2-5",
+    "react-konva": "^18.1.0",
     "react-modal": "^3.14.4",
     "react-papaparse": "^4.0.1",
-    "react-redux": "^8.0.4",
+    "react-redux": "^8.1.3",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.2.2",
     "react-tag-input": "^6.8.0",
     "redux": "^4.2.0",
