@@ -1,10 +1,7 @@
-// import React from "react";
-// import HeroComponent from "../hero/HeroComponent.js";
+import HeroComponent from "../hero/HeroComponent.js";
 
-// export default function HeroTabContent(props) {
-//   return (
-//     <div hidden={props.value !== props.index}>
-//       <HeroComponent id={props.value} />
-//     </div>
-//   );
-// }
+export default function HeroTabContent(props) {
+  return (
+      <HeroComponent id={props.id} />
+  );
+}
