@@ -1,7 +1,7 @@
-import { createStore } from "redux";
-import rootReducer from "./reducers/index";
-import { composeWithDevTools } from "redux-devtools-extension";
-import { loadState, loadBuilds } from "./localStorage";
-const persistedState = loadState();
+// import { createStore } from "redux";
+// import rootReducer from "./reducers/index";
+// import { composeWithDevTools } from "redux-devtools-extension";
+// import { loadState, loadBuilds } from "./localStorage";
+// const persistedState = loadState();
 
-export const store = createStore(rootReducer, persistedState, composeWithDevTools());
+// export const store = createStore(rootReducer, persistedState, composeWithDevTools());
